+++ Tests/Unit/Configuration/WidgetsProviderTest.php
@@ -14,6 +14,7 @@ namespace Brotkrueml\MatomoWidgets\Tests\Unit\Configuration;
 use Brotkrueml\MatomoWidgets\Configuration\WidgetsProvider;
 use Brotkrueml\MatomoWidgets\Extension;
 use PHPUnit\Framework\TestCase;
+use TYPO3\CMS\Core\Information\Typo3Version;
 
 class WidgetsProviderTest extends TestCase
 {
@@ -54,8 +55,12 @@ class WidgetsProviderTest extends TestCase
     /**
      * @test
      */
-    public function getItemsForTca(): void
+    public function getItemsForTcaInV11(): void
     {
+        if ((new Typo3Version())->getMajorVersion() > 11) {
+            self::markTestSkipped('Only relevant for TYPO3 v11');
+        }
+
         $actual = $this->subject->getItemsForTca();
 
         self::assertContains([Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.visitsSummary.actionsPerDay.title', 'actionsPerDay'], $actual);
@@ -75,5 +80,86 @@ class WidgetsProviderTest extends TestCase
         self::assertContains([Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.actions.siteSearchNoResultKeywords.title', 'siteSearchNoResultKeywords'], $actual);
         self::assertContains([Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.visitsSummary.visitsPerDay.title', 'visitsPerDay'], $actual);
         self::assertContains([Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.visitsSummary.visitsPerMonth.title', 'visitsPerMonth'], $actual);
+    }
+
+    /**
+     * @test
+     */
+    public function getItemsForTca(): void
+    {
+        if ((new Typo3Version())->getMajorVersion() < 12) {
+            self::markTestSkipped('Only relevant for TYPO3 v12+');
+        }
+
+        $actual = $this->subject->getItemsForTca();
+
+        self::assertContains([
+            'label' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.visitsSummary.actionsPerDay.title',
+            'value' => 'actionsPerDay',
+        ], $actual);
+        self::assertContains([
+            'label' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.visitsSummary.actionsPerMonth.title',
+            'value' => 'actionsPerMonth',
+        ], $actual);
+        self::assertContains([
+            'label' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.annotations.title',
+            'value' => 'annotations',
+        ], $actual);
+        self::assertContains([
+            'label' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.createAnnotation.title',
+            'value' => 'createAnnotation',
+        ], $actual);
+        self::assertContains([
+            'label' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.visitsSummary.bounceRate.title',
+            'value' => 'bounceRate',
+        ], $actual);
+        self::assertContains([
+            'label' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.referrers.campaigns.title',
+            'value' => 'campaigns',
+        ], $actual);
+        self::assertContains([
+            'label' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.contents.contentNames.title',
+            'value' => 'contentNames',
+        ], $actual);
+        self::assertContains([
+            'label' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.contents.contentPieces.title',
+            'value' => 'contentPieces',
+        ], $actual);
+        self::assertContains([
+            'label' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.userCountry.country.title',
+            'value' => 'countries',
+        ], $actual);
+        self::assertContains([
+            'label' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.events.javaScriptErrors.title',
+            'value' => 'javaScriptErrors',
+        ], $actual);
+        self::assertContains([
+            'label' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.linkMatomo.title',
+            'value' => 'linkMatomo',
+        ], $actual);
+        self::assertContains([
+            'label' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.devicesDetection.osFamilies.title',
+            'value' => 'osFamilies',
+        ], $actual);
+        self::assertContains([
+            'label' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.actions.pagesNotFound.title',
+            'value' => 'pagesNotFound',
+        ], $actual);
+        self::assertContains([
+            'label' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.actions.siteSearchKeywords.title',
+            'value' => 'siteSearchKeywords',
+        ], $actual);
+        self::assertContains([
+            'label' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.actions.siteSearchNoResultKeywords.title',
+            'value' => 'siteSearchNoResultKeywords',
+        ], $actual);
+        self::assertContains([
+            'label' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.visitsSummary.visitsPerDay.title',
+            'value' => 'visitsPerDay',
+        ], $actual);
+        self::assertContains([
+            'label' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.visitsSummary.visitsPerMonth.title',
+            'value' => 'visitsPerMonth',
+        ], $actual);
     }
 }